+++ tests/testthat/test_base_getFeatureImportance.R
@@ -29,7 +29,7 @@ test_that("getFeatureImportance", {
   expect_equal(colnames(feat.imp), mod$features)
 
   #wrapped learner
-  lrn = makeFilterWrapper(makeLearner("regr.gbm"), fw.method = "information.gain", fw.abs = 2)
+  lrn = makeFilterWrapper(makeLearner("regr.gbm"), fw.method = "FSelectorRcpp.infogain", fw.abs = 2)
   mod = train(lrn, regr.task)
   feat.imp = getFeatureImportance(mod)$res
   expect_data_frame(feat.imp, types = rep("numeric", getTaskNFeats(regr.task)),